fix: bigint를 tinyint로 수정했습니다
lotto테이블의 6개 번호 컬럼을 Byte 형식으로 바꾸고 서비스쪽도 에러가 없도록 수정하였습니다
로컬DB에서 정상으로 작동하는것을 확인하였으니 merge 후 lotto테이블을 드랍 한번만 해주세요

issue: #25

+++ src/main/java/com/lotto/lotto_simulator/controller/requestDto/LottoDto.java
@@ -10,16 +10,16 @@ import lombok.Setter;
 @NoArgsConstructor
 public class LottoDto {
 
-    private Long firstNum;
-    private Long secondNum;
-    private Long thirdNum;
-    private Long fourthNum;
-    private Long fifthNum;
-    private Long sixthNum;
+    private Byte firstNum;
+    private Byte secondNum;
+    private Byte thirdNum;
+    private Byte fourthNum;
+    private Byte fifthNum;
+    private Byte sixthNum;
 
     private String store;
     @QueryProjection
-    public LottoDto(Long firstNum, Long secondNum, Long thirdNum, Long fourthNum, Long fifthNum, Long sixthNum) {
+    public LottoDto(Byte firstNum, Byte secondNum, Byte thirdNum, Byte fourthNum, Byte fifthNum, Byte sixthNum) {
         this.firstNum = firstNum;
         this.secondNum = secondNum;
         this.thirdNum = thirdNum;
